==> Analysis & Scraping projects/Tripadvisor project/Tripadvisor Project/1_ParsingData_The Cities With the Highest Rated Independent Coffee Shops.ipynb <==
import numpy as np
import pandas as pd
import requests
from bs4 import BeautifulSoup
import json
import os
import time
import random

# Read scraped data

## Listings

data = pd.read_csv('data.csv')
data

## Shops


ratings_breakdowns = pd.DataFrame(columns=['locationId', 'Excellent', 'Very_Good', 
                                           'Average', 'Poor', 'Terrible'])

json_files = [f for f in os.listdir(folder_path) if f.endswith('.json')]

for file in json_files:
    with open(os.path.join(folder_path, file), 'r') as json_file:
        json_data = json.load(json_file)
        locationId = json_data[0]['data']['locations'][0]['locationId']
        Excellent = json_data[0]['data']['locations'][0]['reviewAggregations']['ratingCounts'][4]
        Very_Good = json_data[0]['data']['locations'][0]['reviewAggregations']['ratingCounts'][3]
        Average = json_data[0]['data']['locations'][0]['reviewAggregations']['ratingCounts'][2]
        Poor = json_data[0]['data']['locations'][0]['reviewAggregations']['ratingCounts'][1]
        Terrible = json_data[0]['data']['locations'][0]['reviewAggregations']['ratingCounts'][0]
    
    # Append to the DataFrame
        ratings_breakdowns = ratings_breakdowns.append({
                'locationId': locationId,
                'Excellent': Excellent,
                'Very_Good': Very_Good,
                'Average' : Average,
                'Poor': Poor,
                'Terrible': Terrible
                }, ignore_index=True)

ratings_breakdowns

data = data.merge(ratings_breakdowns, on=['locationId'], how='left')

data

len(data.City.unique())

# Calculate the weighted average score across all independent coffee shops in each city

# Assign numeric ratings
ratings = {
    'Excellent': 5,
    'Very_Good': 4,
    'Average': 3,
    'Poor': 2,
    'Terrible': 1
}

data['total_score'] = (data['Excellent'] * ratings['Excellent'] +
                     data['Very_Good'] * ratings['Very_Good'] +
                     data['Average'] * ratings['Average'] +
                     data['Poor'] * ratings['Poor'] +
                     data['Terrible'] * ratings['Terrible'])

# Calculate the weighted average rating for each shop
data['weighted_average'] = data['total_score'] / data['reviews']

# Calculate the weighted average across coffee shops in each city and state, and count the number of coffee shops
city_state_weighted_avg = data[~pd.isna(data['weighted_average'])].groupby(['City', 'State']).apply(
    lambda x: pd.Series({
        'city_state_weighted_average': (x['weighted_average'] * x['reviews']).sum() / x['reviews'].sum(),
        'coffee_shops_number': x.shape[0]
    })
).reset_index()

city_state_weighted_avg.sort_values(by='city_state_weighted_average', ascending=False)

city_state_weighted_avg['coffee_shops_number'].describe()

#df = city_state_weighted_avg[city_state_weighted_avg['coffee_shops_number'] >=10] 

#df

data.to_csv('data_after_parsing.csv', index=False)
city_state_weighted_avg.to_csv('city_state_weighted_avg.csv', index=False)
