==> Analysis & Scraping projects/Tripadvisor project/Tripadvisor Project/2_FinallData_The Cities With the Highest Rated Independent Coffee Shops.ipynb <==
import numpy as np
import pandas as pd

# Finalise the data into tabs

data = pd.read_csv('data_after_parsing.csv')
city_state_weighted_avg =pd.read_csv('city_state_weighted_avg.csv')

## City Ranking

City Ranking: rank all cities based on the weighted average across all shops.


City_Ranking = city_state_weighted_avg.sort_values(by='city_state_weighted_average', ascending=False)

City_Ranking

## Top Shops

Top Shops: isolate the top 10 independent shops with the highest rating in the full dataset.


data['reviews'].describe()

Top_Shops = data[data.reviews >=40].sort_values(by=[ 'weighted_average'], ascending=False).head(10)

Top_Shops

## States Shops

States Shops: isolate the highest rated shop in every state

filtered_df = data.copy()
filtered_df['weighted_average'] = pd.to_numeric(filtered_df['weighted_average'], errors='coerce')
filtered_df = filtered_df.dropna(subset=['weighted_average'])

# Filter shops with at least 40 reviews
filtered_df = data[data['reviews'] >= 40]
filtered_df['weighted_average'] = filtered_df['weighted_average'].astype(float)


# Group by 'State' and find the shop with the highest weighted average rating in each state
States_Shops = filtered_df.loc[filtered_df.groupby('State')['weighted_average'].idxmax()]


States_Shops

# Create an Excel file

# Create an Excel writer object

writer = pd.ExcelWriter('The Cities With the Highest Rated Independent Coffee Shops.xlsx')

City_Ranking.to_excel(writer, sheet_name='City_Ranking', index=False)
Top_Shops.to_excel(writer, sheet_name='Top_Shops', index=False)
States_Shops.to_excel(writer, sheet_name='States_Shops', index=False)
data.to_excel(writer, sheet_name='Raw_Data', index=False)

# Save the Excel file
writer.save()
